Synchronizing with https://svn.forge.osor.eu/svn/gvsig-desktop/, commit #34632
git-svn-id: https://gvsigce.svn.sourceforge.net/svnroot/gvsigce/trunk@548 d80a0976-42e5-4521-8f91-9f5439ad514b

+++ libraries/libRaster/src/org/gvsig/raster/buffer/RasterBuffer.java
@@ -118,8 +118,8 @@ public abstract class RasterBuffer implements IBuffer {
 		}
 			
 		if(cacheOn){
-			long size = (RasterUtilities.getBytesFromRasterBufType(dataType) * width * height * bandNr) / 1024;
-			long ms1 = RasterLibrary.cacheSize * 1024;
+			long size = (long)((long)RasterUtilities.getBytesFromRasterBufType(dataType) * (long)width * (long)height * (long)bandNr) / 1024L;
+			long ms1 = RasterLibrary.cacheSize * 1024L;
 			if(size <= ms1)
 				return new RasterMemoryBuffer(dataType, width, height, bandNr, malloc);
 			else
